tables.Filters
By executing the code of Cluster_Ensembles, I found out that "tables.filters" gave me an error: 'module' object not callable. I'm using the most recent version of tables installed from pip with --upgrade option. The problem was solved by using the capital letter "F" in filters call. But please, check if this modification has to be done, as maybe mine was a particular use-case. Since it worked out for me, I just wanted to point that out in case it might be useful. Regards, Fabio Pierazzi